+++ src/main/java/roman/pidkostelnyi/victoriaarmario/service/SubcategoryService.java
@@ -1,9 +1,8 @@
 package roman.pidkostelnyi.victoriaarmario.service;
 
-import lombok.Getter;
-import lombok.Setter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import roman.pidkostelnyi.victoriaarmario.dto.request.SubcategoryRequest;
 import roman.pidkostelnyi.victoriaarmario.dto.response.SubcategoryResponse;
@@ -12,11 +11,8 @@ import roman.pidkostelnyi.victoriaarmario.repository.SubcategoryRepository;
 import roman.pidkostelnyi.victoriaarmario.tool.FileTool;
 
 import java.io.IOException;
-import java.nio.file.Paths;
 import java.util.List;
 import java.util.stream.Collectors;
-
-import static roman.pidkostelnyi.victoriaarmario.tool.Constants.USER_HOME;
 
 @Service
 public class SubcategoryService {
@@ -38,8 +34,8 @@ public class SubcategoryService {
         subcategoryRepository.save(subcategoryRequestToSubcategory(null, request));
     }
 
-    public List<SubcategoryResponse> findAllByCategoryId(Long categoryId) {
-        return subcategoryRepository.findAllByCategoryId(categoryId).stream().map(SubcategoryResponse::new).collect(Collectors.toList());
+    public List<SubcategoryResponse> findAll() {
+        return subcategoryRepository.findAll(Sort.by(Sort.Direction.ASC, "category.name")).stream().map(SubcategoryResponse::new).collect(Collectors.toList());
     }
 
     public void update(Long id, SubcategoryRequest request) throws IOException {
@@ -48,8 +44,12 @@ public class SubcategoryService {
 
     public void delete(Long id) {
         Subcategory subcategory = findOne(id);
-        subcategoryRepository.delete(subcategory);
-        Paths.get(System.getProperty(USER_HOME), imgDirectory, subcategory.getImage()).toFile().delete();
+        if (subcategory.getProducts().isEmpty()) {
+            subcategoryRepository.delete(subcategory);
+            fileTool.deleteFile(imgDirectory, subcategory.getImage());
+        } else {
+            throw new IllegalArgumentException("Subcategory cannot be deleted");
+        }
     }
 
     public Subcategory findOne(Long id) {
